requirements.txt
torch>=1.12.0
torchvision>=0.13.0
einops>=0.4.1
transformers>=4.21.0
numpy>=1.21.0
Pillow>=9.0.0
tqdm>=4.62.0
opencv-python>=4.5.0
pandas>=1.3.0
matplotlib>=3.4.0
streamlit>=1.10.0